Split actions.ts file into categories
