+++ src/modules/BuilderPlace/apiResponses.ts
@@ -8,3 +8,29 @@ export const ALREADY_HAVE_PROFILE = 'You already have a profile';
 export const COLLABORATOR_NOT_FOUND = 'Collaborator not found';
 export const USER_PROFILE_NOT_VERIFIED = 'The user profile is not verified';
 export const ERROR_ADDING_COLLABORATOR = 'Error adding collaborator';
+export const ERROR_REMOVING_COLLABORATOR = 'Error removing collaborator';
+export const ERROR_DELETING_BUILDERPLACE = 'Error deleting the BuilderPlace';
+export const ERROR_UPDATING_BUILDERPLACE = 'Error updating the BuilderPlace';
+export const ERROR_CREATING_BUILDERPLACE = 'Error creating new builderPlace';
+export const ERROR_FETCHING_BUILDERPLACE = 'Error fetching the builderPlace';
+export const ERROR_FETCHING_BUILDERPLACE_OWNER = 'Error fetching the builderPlace owner';
+export const ERROR_UPDATING_DOMAIN = 'Error updating domain';
+export const ERROR_SETTING_BUILDERPLACE_OWNER = 'Error setting builderPlace owner';
+export const ERROR_SETTING_USER_OWNER = 'Error setting User owner';
+export const ERROR_REMOVING_USER_OWNER = 'Error removing User owner';
+export const ERROR_REMOVING_BUILDERPLACE_OWNER = 'Error removing builderPlace owner';
+export const ERROR_REMOVING_USER_ADDRESS = 'Error removing builderPlace owner';
+export const ERROR_VALIDATING_BUILDERPLACE = 'Error validating BuilderPlace';
+export const ERROR_VALIDATING_USER = 'Error validating User';
+export const ERROR_CREATING_WORKER_PROFILE = 'Error creating new Worker Profile';
+export const ERROR_CREATING_HIRER_PROFILE = 'Error creating new Hirer Profile';
+export const ERROR_UPDATING_HIRER_PROFILE = 'Error updating Hirer Profile';
+export const ERROR_UPDATING_WORKER_PROFILE = 'Error updating Worker Profile';
+export const ERROR_FETCHING_USER = 'Error fetching User';
+export const TRANSACTION_LIMIT_REACHED = 'Transaction limit reached for the week';
+export const ERROR_INCREMENTING_TRANSACTION_COUNTER = 'Error incrementing transaction counter';
+export const ERROR_CHECKING_TRANSACTION_COUNTER = 'Error checking transaction counter';
+export const DOMAIN_CONTAINS_BUILDER_PLACE = 'Domain contains builder.place';
+export const INVALID_CUSTOM_DOMAIN = 'Invalid custom domain';
+export const ERROR_REMOVING_CUSTOM_DOMAIN = 'Error removing custom domain';
+export const COLLABORATOR_ALREADY_EXISTS = 'Collaborator already exists';
