+++ src/pages/api/domain/[slug]/get-builder-place.ts
@@ -1,14 +1,18 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { getBuilderPlaceByDomain } from '../../../../modules/BuilderPlace/actions';
+import { getBuilderPlaceByDomain } from '../../../../modules/BuilderPlace/actions/builderPlace';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
-  const { slug: domain } = req.query;
+  try {
+    const { slug: domain } = req.query;
 
-  if (!domain) {
-    return res.status(500).json({ error: 'No domain provided.' });
+    if (!domain) {
+      return res.status(500).json({ error: 'No domain provided.' });
+    }
+
+    const result = await getBuilderPlaceByDomain(domain as string);
+
+    return res.json(result);
+  } catch (error: any) {
+    res.status(400).json({ error: error.message });
   }
-
-  const result = await getBuilderPlaceByDomain(domain as string);
-
-  return res.json(result);
 }
