+++ src/pages/api/domain/create-builder-place.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { CreateBuilderPlaceProps } from '../../../modules/BuilderPlace/types';
-import { createBuilderPlace } from '../../../modules/BuilderPlace/actions';
+import { createBuilderPlace } from '../../../modules/BuilderPlace/actions/builderPlace';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === 'POST') {
@@ -13,9 +13,7 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
 
       res.status(200).json({ message: result.message, id: result.id });
     } catch (error: any) {
-      //TODO check error sending, as of now we receive an empty object
-      console.log('error in controller', error);
-      res.status(400).json({ error: { ...error } });
+      res.status(400).json({ error: error.message });
     }
   } else {
     res.status(405).json({ error: 'Method not allowed' });
