+++ src/pages/api/domain/create-hirer-profile.ts
@@ -1,10 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { CreateHirerProfileProps } from '../../../modules/BuilderPlace/types';
-import {
-  createHirerProfile,
-  getUserByEmail,
-  updateHirerProfile,
-} from '../../../modules/BuilderPlace/actions';
+import { getUserByEmail } from '../../../modules/BuilderPlace/actions/email';
+import { createHirerProfile, updateHirerProfile } from '../../../modules/BuilderPlace/actions/user';
 
 //TODO make it one API to create either worker or hirer ? Or update an existing User when declaring him collaborator ?
 // Or pass as param if the guy is a hirer or worker ?
@@ -35,7 +32,7 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
 
       res.status(200).json({ message: result.message, id: result.id });
     } catch (error: any) {
-      res.status(400).json({ error: error });
+      res.status(400).json({ error: error.message });
     }
   } else {
     res.status(405).json({ error: 'Method not allowed' });
