+++ src/pages/api/domain/delete-builder-place.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { deleteBuilderPlace } from '../../../modules/BuilderPlace/actions';
 import { checkSignature } from '../utils/domain';
 import { DeleteBuilderPlace } from '../../../modules/BuilderPlace/types';
+import { deleteBuilderPlace } from '../../../modules/BuilderPlace/actions/builderPlace';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === 'DELETE') {
@@ -15,7 +15,7 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
       const result = await deleteBuilderPlace(body.id);
       res.status(200).json({ message: result.message });
     } catch (error: any) {
-      res.status(400).json({ error: error });
+      res.status(400).json({ error: error.message });
     }
   } else {
     res.status(405).json({ error: 'Method not allowed' });
