+++ src/pages/api/domain/get-builder-place-from-owner.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { getBuilderPlaceByOwnerId } from '../../../modules/BuilderPlace/actions';
+import { getBuilderPlaceByOwnerId } from '../../../modules/BuilderPlace/actions/builderPlace';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
   const body: any = req.body;
