+++ src/pages/api/domain/get-worker-profile-by-owner-id.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { getUserByTalentLayerId } from '../../../modules/BuilderPlace/actions';
+import { getUserByTalentLayerId } from '../../../modules/BuilderPlace/actions/user';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
   const body: any = req.body;
