+++ src/pages/api/utils/sendgrid.ts
@@ -1,6 +1,6 @@
 import * as sgMail from '@sendgrid/mail';
 import { renderValidationMail } from './renderValidationEmail';
-import { getBuilderPlaceByDomain } from '../../../modules/BuilderPlace/actions';
+import { getBuilderPlaceByDomain } from '../../../modules/BuilderPlace/actions/builderPlace';
 
 const SENDERS_EMAIL = process.env.NEXT_PRIVATE_SENDGRID_VERIFIED_SENDER;
 const APIKEY = process.env.NEXT_PRIVATE_SENDGRID_API_KEY;
